fix: address PR bugbot findings
- Consolidate Storybook config: .storybook/main.ts now has the correct
  sveltekit framework + installed addons. Removed stale .storybook/main.js
  (referenced uninstalled addon-links/onboarding/interactions and
  @storybook/svelte-vite framework) and webapp/main.ts (correct content
  but wrong location — Storybook only reads .storybook/)
- Gitignore local secrets: providers.config.json and leo.config.json
  contain OAuth client_id/client_secret; README instructs devs to create
  them at webapp root, so they must not be committable

Task: ES-2951
