+++ webapp/.gitignore
@@ -15,6 +15,10 @@ sst-env.d.ts
 !.env.example
 !.env.test
 
+# Local config with secrets — see README.md
+providers.config.json
+leo.config.json
+
 # OS
 .DS_Store
 Thumbs.db
