+++ webapp/.storybook/main.ts
@@ -1,0 +1,19 @@
+import type { StorybookConfig } from '@storybook/sveltekit';
+
+const config: StorybookConfig = {
+  stories: [
+    '../src/**/*.mdx',
+    '../src/**/*.stories.@(js|ts|svelte)',
+  ],
+  addons: [
+    '@storybook/addon-essentials',
+    '@storybook/addon-svelte-csf',
+    '@chromatic-com/storybook',
+    '@storybook/experimental-addon-test',
+  ],
+  framework: {
+    name: '@storybook/sveltekit',
+    options: {},
+  },
+};
+export default config;
